Rewrite server as a connect middleware
