+++ lib/server.js
@@ -1,7 +1,8 @@
 var Watcher = require('./watcher')
-var path = require('path')
-var hapi = require('hapi')
+var middleware = require('./middleware')
+var http = require('http')
 var tinylr = require('tiny-lr')
+var connect = require('connect')
 
 exports.serve = serve
 function serve (builder, options) {
@@ -11,54 +12,9 @@ function serve (builder, options) {
 
   var watcher = new Watcher(builder)
 
-  var server = hapi.createServer(options.host, parseInt(options.port, 10), {
-    views: {
-      engines: {
-        html: 'handlebars'
-      },
-      path: path.join(__dirname, '../templates')
-    },
-    state: {
-      cookies: {
-        // If we ever need to parse cookies, be sure to set
-        // failAction/strictHeader to be tolerant
-        parse: false
-      }
-    }
-  })
+  var app = connect().use(middleware(watcher))
 
-  var currentBuildDirectory
-
-  server.route({
-    method: 'GET',
-    path: '/{path*}',
-    handler: {
-      directory: {
-        path: function (request) {
-          if (!currentBuildDirectory) throw new Error('Expected currentBuildDirectory to be set')
-          return currentBuildDirectory
-        }
-      }
-    }
-  })
-
-  server.ext('onRequest', function (request, reply) {
-    // If a build is currently running, this will defer until it has finished
-    watcher.then(function(directory) {
-      currentBuildDirectory = directory
-      reply()
-    }, function(buildError) {
-      currentBuildDirectory = null
-      var context = {
-        message: buildError.message,
-        file: buildError.file,
-        line: buildError.line,
-        column: buildError.column,
-        stack: buildError.stack
-      }
-      reply.view('error', context).code(500)
-    })
-  })
+  var server = http.createServer(app)
 
   // We register these so the 'exit' handler removing temp dirs is called
   process.on('SIGINT', function () {
@@ -95,5 +51,5 @@ function serve (builder, options) {
     liveReload()
   })
 
-  server.start()
+  server.listen(parseInt(options.port, 10), options.host)
 }
