+++ lib/watcher.js
@@ -26,5 +26,5 @@ Watcher.prototype.check = function() {
 }
 
 Watcher.prototype.then = function(success, fail) {
-  this.current.then(success, fail)
+  return this.current.then(success, fail)
 }
