+++ package.json
@@ -10,7 +10,6 @@
     "url": "https://github.com/joliss/broccoli"
   },
   "dependencies": {
-    "hapi": "^2.6.0",
     "glob": "~3.2.6",
     "mkdirp": "^0.3.5",
     "handlebars": "^1.3.0",
@@ -20,7 +19,9 @@
     "rsvp": "^3.0.3",
     "ncp": "^0.5.0",
     "quick-temp": "^0.1.0",
-    "bower-config": "^0.5.0"
+    "bower-config": "^0.5.0",
+    "connect": "~2.14.1",
+    "send": "~0.2.0"
   },
   "devDependencies": {
     "jshint": "~2.3.0"
